Refactor: Split executor/tests.rs into focused test modules
Split monolithic 876-line test file into three focused modules:
- tests/expression_eval.rs (99 lines) - Expression evaluator tests
- tests/select_executor.rs (654 lines) - SELECT execution tests
- tests/limit_offset.rs (122 lines) - Pagination tests

Benefits:
- Improved navigation and organization
- Clear separation of test concerns
- Easier to locate and maintain specific test categories
- 25% reduction in largest test file size (876 → 654 lines)

Note: Pre-existing compilation error in select_executor.rs:471
(SqlValue: Ord trait bound) exists in main branch - not introduced
by this refactoring. File splitting is structurally correct.

Closes #17
